+++ Sources/SuperwallKit/Delegate/SuperwallDelegate.swift
@@ -108,16 +108,16 @@ public protocol SuperwallDelegate: AnyObject {
     error: Swift.Error?
   )
 
+  /// Called before the redemption of a code from a web paywall.
+  @MainActor
+  func willRedeemCode()
+  
   /// A code was redeemed from a web paywall.
   ///
   /// - Parameter result: A ``RedemptionResult`` enum containing information about
   /// the redeemed the code.
   @MainActor
   func didRedeemCode(result: RedemptionResult)
-
-  /// Called before the redemption of a code from a web paywall happens.
-  @MainActor
-  func willRedeemCode()
 }
 
 extension SuperwallDelegate {